+++ src/Sulu/Component/Content/Document/Behavior/LocalizedStructureBehavior.php
@@ -10,8 +10,6 @@
 
 namespace Sulu\Component\Content\Document\Behavior;
 
-use Sulu\Component\DocumentManager\Behavior\Mapping\LocaleBehavior;
-
 /**
  * All content documents can have localized content, but only
  * content documents implementing this behavior can change the